// src/interfaces/formNewClient.ts
export interface INewClientForm {
  // {
  //     "nombre": "Ana",
  //     "apellidoPaterno": "Rodriguez",
  //     "apellidoMaterno": "Rodriguez",
  //     "fechaNacimiento": "1990-01-01",
  //     "rut": 9878745,
  //     "dv": "k",
  //     "sexo": "M",
  //     "telefono": 97878451,
  //     "direccion": "calle 1",
  //     "email": "ana@ana123.cl"

  nombre: string;
  apellidoPaterno: string;
  apellidoMaterno: string;
  fechaNacimiento: string;
  rut: number | string;
  dv: string;
  sexo: string;
  telefono: number | string;
  direccion: string;
  email: string;
}